+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 `Unreleased`_
 =============
 
+- |Feature| Add support for LightGBM models in :class:`~empulse.models.CSBoostClassifier` and
+  :class:`~empulse.models.B2BoostClassifier`
+- |API| :func:`~empulse.metrics.make_objective_churn` now takes a ``model`` argument to calculate the churn objective
+  for either XGBoost or LightGBM models.
+- |API| XGBoost is now an optional dependency together with LightGBM. To install the package with
+  XGBoost and LightGBM support, use the following command: ``pip install empulse[optional]``
 - |Feature| Add parameter validation for all models and samplers
 - |API| Make all arguments of dataset loaders keyword-only
 - |Fix| Update the descriptions attached to each dataset to match information found in the user guide